Fix floor moving button
