+++ Submerged/SubmergedPlugin.cs
@@ -40,7 +40,9 @@ public sealed partial class SubmergedPlugin : BasePlugin
         ResourceManager.CacheSprite("CreateGameBG");
         ResourceManager.CacheSprite("FilterIcon");
         ResourceManager.CacheSprite("FloorDown");
+        ResourceManager.CacheSprite("FloorDownHover");
         ResourceManager.CacheSprite("FloorUp");
+        ResourceManager.CacheSprite("FloorUpHover");
         ResourceManager.CacheSprite("Logo", 400);
         ResourceManager.CacheSprite("Logo", 250, "OptionsLogo");
         ResourceManager.CacheSprite("OptionsBG");
